requirements.txt
# ============================================================
# requirements.txt – alle Abhängigkeiten für den RAG Chatbot
# ============================================================
# Installation: pip install -r requirements.txt

# LlamaIndex Core – alle Pakete müssen dieselbe Core-Version nutzen!
# llama-index-llms-anthropic 0.6.x braucht core>=0.12.36,<0.13
llama-index-core>=0.12.36,<0.13

# LlamaIndex ChromaDB Integration
llama-index-vector-stores-chroma>=0.4.0

# LlamaIndex Anthropic Integration – Claude als LLM für die Antworten
llama-index-llms-anthropic==0.6.19

# LlamaIndex HuggingFace Embeddings – kostenlose lokale Embeddings (kein API-Key nötig!)
# Läuft komplett lokal auf deinem Mac – keine API-Kosten für Embeddings
llama-index-embeddings-huggingface>=0.5.0

# sentence-transformers – wird von HuggingFace Embeddings benötigt
sentence-transformers==3.0.1

# ChromaDB – speichert unsere Dokument-Embeddings lokal auf der Festplatte
# ChromaDB 1.x hat Rust-Bindings die Probleme machen – bei 0.6.x bleiben
chromadb==0.6.3

# Streamlit – baut das Web-UI mit wenigen Zeilen Python
streamlit==1.36.0

# python-dotenv – liest die .env Datei (API-Keys sicher laden)
python-dotenv==1.0.1

# pypdf – liest PDF-Dateien aus (LlamaIndex braucht das intern)
pypdf>=5.1.0
